feature(Lotto-step3): 피드백 내용을 반영하다
- winningResult 에 대한 테스트 케이스를 이동하다
